Renamed annotated region package to annotated interval.
