+++ src/main/resources/org/broadinstitute/hellbender/tools/copynumber/utils/annotatedinterval/annotated_region_default.config
