+++ src/test/java/org/broadinstitute/hellbender/tools/copynumber/utils/annotatedinterval/AnnotatedIntervalUnitTest.java
@@ -1,4 +1,4 @@
-package org.broadinstitute.hellbender.tools.copynumber.utils.annotatedregion;
+package org.broadinstitute.hellbender.tools.copynumber.utils.annotatedinterval;
 
 import com.google.common.collect.ImmutableSortedMap;
 import org.broadinstitute.hellbender.GATKBaseTest;
