+++ src/test/resources/org/broadinstitute/hellbender/tools/copynumber/utils/annotatedinterval/annotated-interval-alt-samheader.seg
