+++ src/test/resources/org/broadinstitute/hellbender/tools/copynumber/utils/annotatedinterval/annotated-interval-collection-index-cols.config
