+++ src/test/resources/org/broadinstitute/hellbender/tools/copynumber/utils/annotatedinterval/annotated-interval-many-columns.seg
