+++ src/test/resources/org/broadinstitute/hellbender/tools/copynumber/utils/annotatedinterval/old-header.config
