+++ src/test/resources/org/broadinstitute/hellbender/tools/copynumber/utils/annotatedinterval/simple-annotated-interval-writer-replacement-header-comments.seg
